+++ src/web/src/Mux.js
@@ -1,0 +1,9 @@
+// import JS
+import { MuxElementSelectInput } from './js/MuxElementSelectInput.js';
+
+document.addEventListener("DOMContentLoaded", function() {
+    window.RocketPark = window.RocketPark || {};
+    window.RocketPark.Mux = window.RocketPark.Mux || {};
+
+    window.RocketPark.Mux.MuxElementSelectInput = MuxElementSelectInput;
+});